+++ public/blogContent/blog005.md
@@ -1,0 +1,33 @@
+# 2017 Review
+
+It was an interesting Year .....
+
+The year started off well, with the release of the first **Nano Guide**, self published to the Amazon Store.
+
+>**Atom - Up & Running**
+
+I also update my tech knowledge with undertaking some online training.
+
+Some subjects covered:
+
+* Modern Web Devlopment
+  * Including HTML / CSS / JavaScript / Bootstrap / MongoDB and associated tooling.
+* ReactJS
+* Axway Appcelerator (Updating for latest releases).
+
+Work wise, started looking for a new full time role, without success but I am sure the correct role will appear.
+
+I undertook a few freelance projects for clients in the US and UK, focusing on two areas, mobile developmet using the Axway Appcelerator tooling, and a new project using ReactJS and Parse.
+
+Over the year I attended a few user groups, Blockchain Nottingham and Ruby Users Sheffield, and one conference GDG London. No speaker slots this year.
+
+The user groups were interesting and it was great to meetup with some old friends and make new ones. The GDG event in London was excellent learning loads about Google developer tooling and came away very interested in Flutter.
+
+I also started running again, and stopping due to injury (we're almost back on this).
+
+My music took a back seat due to a sticky finger, making it impossible to play the Mandola or Tenor guitar (Not that I could play them very well anyway). The good news is I had a minor operation on my hand in November, which has fixed the issue and iis now fully recovered. 2018 should see me get back to learning how to play.
+
+And that was 2017 in brief. 
+
+
+
