+++ public/blogContent/rblog001.md
@@ -2,13 +2,13 @@
 
 I used to be very fit, in my mid teens I swam at County Level, progressing to running cross country and marathons in my late teens and early twenties, and keeping fit (yep full six pack) until about 30.
 
-I have just turned 53 and basically (excluding a few false starts) not been physically active for 20yrs, so really really unfit. The false starts have included the NHS couch to 5k training, online course etc etc, but I just found the effort and control too much to entise me to keep going.
+I have just turned 53 and basically (excluding a few false starts) not been physically active for 20yrs, so really really unfit. The false starts have included the NHS couch to 5k training, online course etc etc, but I found that trying to stick to the routines actually put me off.
 
-I literally woke up the day after my 53rd Birthday and went, I really need to do something about this, but if I start to get fit, and only let people who won't comment if I stop again then it won't work.
+I woke up the day after my 53rd Birthday and went, ***I really need to do something about this***.
 
-Having done so much training in the past I decided I would just apply what I knew. Not push to hard and slowly build up my fitnes. Starting with a sensible distance and walking a lot with some jogging inbetween, then building it up.
+Having done so much training in the past I decided I would just apply what I knew. Not push to hard and slowly build up my fitnes. Starting with a sensible distance and walking a lot with some jogging inbetween, then building it up.  
 
-> The key to this would be going out 3 times a week and being consitant.
+> The key to this would be going out 3 times a week, being consistant and realising it took 20 years to get this unfit, ***There are no quick fixes***
 
 Hence setting my first **GOAL** 
 > **"be able to complete a Park Run by the first weekend in September".**
@@ -20,24 +20,26 @@ When I say complete, my aim is to walk or jog without stopping the full 5k.
 
 So I registered on the **[parkrun website](http://www.parkrun.org.uk/)** and then investigated how people kept motivated. One of the first things I found was **[Strava](https://www.strava.com/)** a useful site to enable me to keep track and set the runs I wanted to do. It enables me to share on instagram, and I have used it to not only keep track for myself, but also let my friends and others (it's a great community) know what I am upto.
 
+I already had a lot of kit, including a fairly unused pair of running shoes (the only really important part of the kit). But I am very self concious of how people see me, so the current running takes place at night.
+
 > **The Journey started with my first training on the *25th July 2017* a few days after my 53rd Birthday.**
 
->>This was a hard session as I was 15st 13 (223 pounds 101.15 kilos) and even the walking bits hurt.    
+>>This was a hard session as I was 15st 13 (223 pounds 101.15 kilos) **totally unfit** and even the walking bits hurt.    
   
 
 #### The First Session 
-![The first Session](./assets/runningblog/firstRun.jpg?style=centerme)
+![The first Session](./assets/runningBlog/firstRun.jpg?style=centerme)
 
 #### The Second Session 
-![The Second Session](./assets/runningblog/secondRun.jpg?style=centerme)
+![The Second Session](./assets/runningBlog/secondRun.jpg?style=centerme)
 
 #### The Third Session 
-![The Third Session](./assets/runningblog/thirdRun.jpg?style=centerme)
+![The Third Session](./assets/runningBlog/thirdRun.jpg?style=centerme)
 
 I am now on week 4 and have moved upto 3.8km walking & jogging combo. I am finding that the jogging parts are getting longer and the walking parts area getting easier. 
 
 #### Week 4 Session 1
-![The Third Session](./assets/runningblog/week4run1.jpg?style=centerme)
+![Week Four Session 1](./assets/runningBlog/week4run1.jpg?style=centerme)
 
 I now feel confident that if I keep the consistency going I will be able to complete my first **Park Run** on schedule, and move on to my next set of **"Fitness Goals"**
