+++ public/blogContent/rblog002.md
@@ -1,0 +1,49 @@
+# My First Parkrun
+
+Six weeks ago I got up and decided to start running again, the first time for over 20 years.
+
+>Old, overweight, and completely unfit. I needed an objective to get and keep going. I set completing my first parkrun (jogging and walking) without stopping on Saturday 2nd September 2017, at Poolsbrook Park, Chesterfield.
+
+I am pleased to say that on Saturday 2nd September 2017 I achieved my objective, in 44 minutes and 37 seconds.
+
+#### Poolsbrook Park Run
+![Poolsbrook Park Run](./assets/runningBlog/stravaroute.jpg?style=centerme)
+
+>During my training for this I have consistently used the Strava mobile app to record and inform people how I am doing.
+
+### The Parkrun
+
+The night before I set out my kit, made sure I had dry clothes incase of rain, set my alarm and setup sat nav... (the park run being over 8 miles from home).
+
+Feeling very apprehensive, as I had not done a 5km distance and not really sure what the run was going to be like.
+
+Well I got up, arrived the sun was shinning and found a group of people looking like they were going running.
+
+![PoolsBrook](./assets/runningBlog/poolsbrook01.png?style=centerme)
+
+![Poolsbrook](./assets/runningBlog/poolsbrookwaiting170902.jpg?style=centerme)
+
+It was well signed, so I knew where I had to go. After the first timers briefing I made my way to the start line, joining the 163 other runners. It being my first time, I headed to the back and quite expected to be enjoying the company of the tail walkers.
+
+![Poolsbrook](./assets/runningBlog/poolsbrookstart170902.jpg?style=centerme)  
+
+Although I combined jogging and walking, I was surprised how I managed, six weeks of training is not a lot but at no point did I think about stopping. It appears when you train in a very hilly area and then run on the flat it's a lot easier.
+
+![Strava Elevation Comparison](./assets/runningBlog/elevation170902.jpg?style=centerme)
+
+I found the event, to be well organised with very friendly people and a nice route. I think the best part was finishing, I knew I had completed my first goal to get back to fitness (still a long way to go), although very hot and very sweaty I received my token, and handed it in with my barcode. Shortly after my results were on the Parkrun website.
+
+![The show and tell](./assets/runningBlog/website.png?style=centerme)
+
+Thank you to everybody who organises the Park Run events and especially those on the Pools Brook for Saturday.
+
+>I will be back.
+
+### Whats Next
+
+I really need to keep this running thing going, and as I used the Parkrun as motivation to start I will be using it as my continuing motivation, with two new goals.
+
+>  **Complete 10 Parkruns by the end of 2017**  
+>  **Be jogging / running (no walking) the Park run by the end of the 2017.**  
+
+I think these are sensible goals, I have been unfit and overweight for over 20 years, and being realistic about my objectives hopefully means I will achieve them. ***(but not to easy).***
