+++ public/blogContent/rblog003.md
@@ -1,0 +1,27 @@
+# Three Months Injured
+
+Been a while since my last post, here is a quick update.
+
+The training was going well, I had got upto running 1k without having to walk, completed 4 park runs, **THEN**
+
+>**INJURY HIT** October hated me running.
+
+My knee started to hurt, and I made the mistake of thinking it was to do with lack of use and needed to be made stronger.   
+
+>**Self diagnosis is dangerous.**
+
+After two weeks of it getting worse, I went to the Doctors, and was told, complete rest and at least 6 weeks for it to recover.
+
+It actually took until the week before Christmas to stop hurting, I then decided to leave it until after Christmas to start to train again.
+
+>**3 Months Injured**
+
+It has meant I didn't achieve my goals and lost the momentum to run. 
+
+It is concerning to get back out there **running**. The thought of getting injured is right at the front of my mind, but I must.
+
+Again my incentive is now to use the Park Run's, Strava and the associated groups of people, to get me remotivated.
+
+The plan is to go for a preliminary run this weekend, seeing how the knee holds up. (When I say run I mean a fast walk and jog).
+
+>***TIME TO GET FIT AGAIN***
