+++ public/blogContent/showpost.json
@@ -4,11 +4,11 @@
   ],
   "posts": [
     {
-      "title":        "The First Few Steps",
-      "file":         "rblog001.md",
-      "date":         "August 17th 2017",
-      "sortDate":     "20170817",
-      "author":        "Trevor Ward"
+      "title":        "2017 Review",
+      "file":         "blog005.md",
+      "date":         "January 10th 2018",
+      "sortDate":     "20180110",
+      "author":       "Trevor Ward"
     }
   ]
 }
