+++ public/content/me.md
@@ -1,80 +1,55 @@
 # Bio
 
-With over 25 years in the IT industry, I have been fortunate to use a multitude of technologies. Being first, and foremost, a developer, currently building enterprise mobile applications using the Appcelerator Toolset and node.js, with knowledge and experience of development methodologies, techniques and team management.
+>**Developing business applications on multiple platforms for 24 years, has enabled me to gain a wealth of knowledge and experience in software development, deployment, enhancement, support, training and team leading, utilising a range of programming languages, technologies, tooling and methodologies.**
 
-Within programming I find the most interesting aspect is the project itself. Although a JavaScript and open source advocate, I do not restrict myself to any one technology and instead take the approach of the project dictates the required solution. This has resulted in a depth of experience in researching, learning and implementing new technologies.
+### Technology
+##### Recent
+JavaScript, ReactJS, NodeJS, Axway Appcelerator Framework, App release process (IOS & Android), Html, CSS, SQL, Grunt, Git, Atom, Bash shell, Markdown, Grunt, Mac, Scrum, Agile development process.
 
-### Technical Skills
+##### Previous
+Shell Script, TDD, Perl, SQL (Oracle & MySQL), Cobol, Unix, Windows, Linux, ‘C’, Sed & Awk, AS400, Dec VAX, Xerox Printers, Ruby on Rails, Adobe Flex, ActionScript, Jasmine, Jenkins, WordPress.
 
-Software Development
-
-+ Mobile Development 7 yrs.
-	+ IOS & Android using the Appcelerator Titanium Framework
-+ JavaScript, HTML, CSS, SQL
-+ ReactJS
-+ NodeJS
-+ Grunt
-+ Git
-+ Atom
-+ Shell Script
-+ TDD
-+ Agile process  
-
-Multi device testing, Full App release process through Apple and Google Stores
-Web Development 12 yrs+,
-Perl, HTML, CSS, SQL (Oracle & MySQL), Web browsers
-Cobol Dev 8 yrs.
-Unix, Mac, Linux
-
-### Other Technologies
-
-These were used for specific projects and roles.
-‘C’, Sed & Awk, AS400, Dec VAX, Xerox Printers, Ruby on Rails, Adobe Flex, ActionScript
-
-### Published Books
-
-##### **Atom Up & Running**
+### Publications
+Atom Up & Running  
 Released January 2017  
+A Nano Guide introducing the Atom editor from Github for use on MacOS  
 
->*A Nano Guide introducing the Atom editor from Github  
-for use on MacOS*  <https://goo.gl/fhu0es>
-
-##### **Appcelerator Titanium ~ Patterns and Best Practices**
+Appcelerator Titanium - Patterns and Best Practices  
 Released: February 2013  
-Co-authored with Boydlee Pollentine  
+Co-authored with Boydlee Pollentine.
 
->*The book takes a step-by-step approach to help you understand CommonJS and Titanium architecture patterns, with easy to follow samples and plenty of in-depth explanations.* <https://goo.gl/NVHTLk>
-
-##### **Augmented Reality ~ Using Appcelerator Titanium**  
+Augmented Reality - Using Appcelerator Titanium  
 Released: October 2012
 
->*This book will show you how to build an Augmented Reality application, which is clean, efficient and usable* <https://goo.gl/HvhZ7n>
+### Current Work
+**7dayTEK Ltd**  
 
+My personal company used to self-publish technical manuals, and undertake freelance work.
 
-### Volunteer
+Working on mobile projects using the Axway Appcelerator tooling, and a large ReactJS application utilising Parse.
 
-##### **Monkey Park CIC**  
-July 2016 – current		Volunteer Director
+### Previous Employment
 
->*Monkey Park is a social enterprise consisting of a Café, Bakery, Bike worshop and co working space. I am one of 6 volunteer directors who aid in running the company.*
+* First Utility Limited  
+* Mobestar Limited  
+  * M & S  
+* Softly Wired Limited  
+  * 3M, Heinz, BP
+* Bit 10 Limited
+  * Jaguar Cars
+* Sleepy Dog Limited  
+* Softly Wired Limited  
+* Orbit UK
+  * IBM, Jaguar Cars, Ordina
+* Safe Computing Limited
+  * BNFL, BT, STW
 
-##### **SAGGA**  
-1989 – Current			Member (www.sagga.org.uk)
+### Activities / Hobbies
+I regularly attend user groups and conferences, including speaking, presenting and co-organising.
 
->*SAGGA is a national Scout and Guide group, who provide service to the Scouting and Guiding movements.*
+##### SAGGA	
+A national support group, providing service to the Scouting and Guiding movements.  
+<http://www.sagga.org.uk>
 
-##### **Scouting**
-
->*Scouting has been a major part of my life since the age of 8. I have attended, organised and ran many activities camps and events. It has enabled me to be resourceful, independent, work as a team and develop personally.*
-
-### User Groups & Conferences
-
->*I have been privileged over the years to be involved in attending, organising and speaking at many technical user groups and conferences.*
-
-##### **ConnectJS ~ TiConf (Europe and US)	~ London JavaScript**
-
->*Videos of my talks can be found on YouTube.*
-
-##### **Spire Digital(co-organiser) ~ Titanium London ~ Flash Midlands ~Sheffield Geeks**
-
->*I have been involved in helping organise these groups, often speaking. I have also been involved in organising Flash camps run in the UK.*
+##### Music					
+Playing the Tenor range of instruments, Tenor Guitar, Mandola, Mando Chello.
