+++ public/content/projects.md
@@ -1,16 +1,15 @@
-***Over the years I have undertaken a few projects, which I can publish or open source. Below are a few I hope are of interest.***
+##### **Over the years I have undertaken a few projects, which I have published or open sourced. Below are a few I hope are of interest.**
 
 ## Atom Up & Running 
 #### A NANO GUIDE
 
 >Atom Up & Running is an introduction to the Atom editor from Github for use on MacOS. It covers the basics of installation, usage, themes, packages and updating.
 >
->**Available in all Amazon stores.**
+
+#### **Currently Unavailable (Update coming soon).**
 >
->###### [US Store $1.51](http://amzn.to/2qTwAZh)
->###### [UK Store £1.21](http://amzn.to/2r1jJ7O)
 >
->> ##### A Nano Guide is:
+> ##### A Nano Guide is:
 >> An eBook, which explains a specific area of a subject, saving having to buy a full manual or trawl the Internet for a series of blog posts. Nano Guides are small, 20 – 30 pages long, detailed, economical and convenient.
      
 ## React Website
