+++ src/components/App.js
@@ -11,7 +11,6 @@ var Projects              = require('./Projects');
 var Me                    = require('./Me');
 var Blog                  = require('./Blog');
 var NotFound              = require('./NotFound');
-var RightSideBar          = require('./RightSideBar');
 var SocialBar             = require('./SocialBar');
 var RunningBlog           = require('./RunningBlog');
 
@@ -20,9 +19,9 @@ class App extends React.Component {
     return (
       <Router history={BrowserHistory}>
         <div>
-          <div className="headingBit"> </div>
+          <div className="spacer40"> </div>
 
-          <div className="container greenBack">
+          <div className="container whiteBack">
             <div className="row headingArea">
               <HeadingSection />
             </div>
@@ -38,12 +37,13 @@ class App extends React.Component {
                   <Route path="/*"                        component={NotFound} status={404} />
                 </Switch>
               </div>
-              <div className="col-sm-1"><RightSideBar /></div>
+              <div className="col-sm-1"></div>
             </div>
           </div>
           <footer className="footerBar">
             <SocialBar />
           </footer>
+          <div className="spacer40"></div>
         </div>
       </Router>
     );
