+++ src/components/StrapLine.js
@@ -3,7 +3,7 @@ var React                 = require('react');
 class StrapLine extends React.Component {
   render () {
     return (
-        <div className="strapline headFontSmall orangeColor">
+        <div className="strapline headFontSmall DGreenColor">
             Husband  ~  Father  ~  Published Author  ~  Coder  ~  Blogger  ~  Geek  ~  JavaScript Junkie
         </div>
     )
