+++ src/styles/index.css
@@ -1,32 +1,36 @@
 :root  {
-  --blue:     #345678;
-  --green:    #F3FFE2;
+  --blue:     #000022;
+  --green:    #EEFFEE;
   --orange:   #EA601A;
   --white:    #FFFFFF;
+  --grey:     #939393;
+  --black:    #000000;
+  --DGreen:   #007711;
+  --links:    #009944;
   --HeadFont: HoboStd;
   --BodyFont: Merriweather;
 };
 @font-face {
   font-family: 'HoboStd', 'Merriweather';
-
   src: url('../fonts/HoboStd.otf');
   src: url('../fonts/Merriweather-Regular.ttf');
 }
 
 body {
-  background: var(--blue);
+  background: var(--black);
   margin: 0;
   padding: 0;
   font-family: var(--BodyFont);
   font-size: 16px;
 }
 h1, h2, h3, h4, h5, h6 {
-  color: var(--blue);
+  color: var(--DGreen);
   font-family: var(--HeadFont);
 }
 a {
-  color: orange;
+  color: var(--links);
   font-size: 16px;
+  font-weight: 600;
 }
 a:hover {
   color: inherit;
@@ -44,20 +48,24 @@ label {
   margin: 0;
   float: left;
   padding-left: 10px;
-  color: var(--blue);
+  color: var(--DGreen);
   font-family: var(--BodyFont);
   font-style: italic;
+  font-weight: 500;
 }
 blockquote {
   margin-left: 25px;
   background-color: inherit;
   padding: 10px 20px 10px 20px;
   font-size: 16px;
-  border-left: 5px solid var(--orange);
+  border-left: 5px solid var(--DGreen);
 }
 img[src$="centerme"] {
   display:block;
   margin: 0 auto;
+}
+.DGreenBack {
+  background-color: var(--DGreen);
 }
 .greenBack {
   background-color: var(--green);
@@ -76,6 +84,12 @@ img[src$="centerme"] {
 }
 .orangeColor {
   color: var(--orange);
+}
+.DGreenColor {
+  color: var(--DGreen);
+}
+.whiteBack {
+  background-color: var(--white);
 }
 .whiteColor {
   color: var(--white);
@@ -101,21 +115,18 @@ img[src$="centerme"] {
 }
 /* The Components */
 .headingArea {
-  padding: 5px 5px 0 5px;
+  padding: 5px 5px 5px 5px;
   height: auto;
-  border-top: 2px solid var(--orange);
-  border-right: 2px solid var(--orange);
-  border-left: 2px solid var(--orange);
-  border-bottom: 2px solid var(--orange);
-  background-image: url("/assets/headerImages/h3.jpg");
+  border: 4px solid var(--DGreen);
+  background-color: var(--black);
   background-repeat: no-repeat;
   background-position: left;
 }
 .bottomArea {
-  padding: 10px 0 25px 0;
-  border-bottom: 2px solid var(--orange);
-  border-right: 2px solid var(--orange);
-  border-left: 2px solid var(--orange);
+  padding: 10px 0 30px 0;
+  border-bottom: 4px solid var(--DGreen);
+  border-right: 4px solid var(--DGreen);
+  border-left: 4px solid var(--DGreen);
   height: auto;
 }
 .footerBar {
@@ -130,10 +141,10 @@ img[src$="centerme"] {
   text-align: center;
 }
 .seperatorLine {
-  margin: 5px 0 5px 0;
+  margin: 5px 0 7px 0;
   border: 0;
-  height: 2px;
-  background: var(--blue);
+  height: 3px;
+  background: var(--DGreen);
 }
 .blankLine {
   margin: 1px 0 1px 0;
@@ -144,17 +155,13 @@ img[src$="centerme"] {
 .headerBar {
   width: 330px;
 }
-.headingBit {
-  height: 20px;
-  width: 100%;
-}
 .menuBar {
   color: var(--white);
   width: 250px;
   padding: 0;
 }
 .blogSelectionBar {
-  height: 40px;
+  height: 45px;
   width: 100%;
   padding: 0;
   margin: 0;
@@ -162,13 +169,14 @@ img[src$="centerme"] {
 .blogSelect {
     float: right;
     margin:  0 10px 0 0;
-    padding: 0 10px 0 10px;
-    border: 2px solid var(--blue);
+    padding: 5px 10px 5px 10px;
+    border: 2px solid var(--DGreen);
     width: auto;
     border-radius: 3px;
     outline: none;
-    background: var(--green);
-    color: var(--blue);
+    background: var(--white);
+    color: var(--DGreen);
+    font-weight: 500;
 }
 .blogInfoBar {
   width: 100%;
@@ -180,16 +188,18 @@ img[src$="centerme"] {
   width: auto;
   float: left;
   padding-left: 10px;
-  color: var(--blue);
+  color: var(--DGreen);
   font-style: italic;
+  font-weight: 500;
 }
 .blogInfoDate {
   width: auto;
   float: right;
   height: auto;
   padding-right: 10px;
-  color: var(--blue);
+  color: var(--DGreen);
   font-style: italic;
+  font-weight: 500;
 }
 .socialItem {
   padding: 0;
@@ -240,8 +250,8 @@ img[src$="centerme"] {
   padding: 7px;
   height:  34px;
   text-decoration: none;
-  border: 2px solid var(--blue);
-  background-color: var(--green);
+  border: 2px solid var(--DGreen);
+  background-color: var(--white);
   border-radius: 12px;
 }
 .iconText {
@@ -262,6 +272,7 @@ img[src$="centerme"] {
 .contentArea {
   color: var(--blue);
   font-family: var(--BodyFont);
+  background-color: var(--white);
 }
 .sideBar {
   color: var(--blue);
@@ -271,7 +282,7 @@ img[src$="centerme"] {
   border: 0;
   height: 2px;
   width: 75%;
-  background: var(--blue);
+  background: var(--DGreen);
 }
 .smallRoundImage {
   padding: 0;
@@ -303,4 +314,7 @@ img[src$="centerme"] {
 }
 .circleRadius {
   border-radius: 50%;
+}
+.spacer40 {
+  height: 40px;
 }
